requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

albumentations==1.3.1
einops==0.7.0
imageio==2.33.1
matplotlib==3.8.2
opencv-python-headless==4.10.0.84
pandas==2.1.4
pycocotools==2.0.7
pypng==0.20220715.0
pytorch-lightning==2.2.5
scikit-learn==1.3.2
torch==2.1.2+cu118
torch-kmeans==0.2.0
torchmetrics==1.6.1
torchvision==0.16.2+cu118
transformers==4.49.0
visdom==0.2.4
timm==0.9.10
kornia
hydra-core